test: share historical attribution upstream fakes
