+++ tests/integration/test_historical_attribution_endpoint.py
@@ -2,7 +2,10 @@ from fastapi.testclient import TestClient
 
 from app.main import app
 from tests.support.app_runtime import override_app_runtime
-from tests.support.returns_series_payloads import build_returns_series_response
+from tests.support.historical_attribution_fakes import (
+    RecordingHistoricalAttributionCoreClient,
+    build_stateful_attribution_returns_client,
+)
 
 
 def _stateless_attribution_payload() -> dict[str, object]:
@@ -196,88 +199,8 @@ def test_historical_attribution_stateless_happy_path() -> None:
 
 
 def test_historical_attribution_stateful_total_risk_happy_path() -> None:
-    class _FakeLotusPerformanceClient:
-        def __init__(self) -> None:
-            self.calls: list[dict[str, object | None]] = []
-
-        async def get_returns_series(
-            self,
-            *,
-            request_payload: dict[str, object],
-            correlation_id: str | None,
-        ) -> dict[str, object]:
-            self.calls.append(
-                {
-                    "request_payload": request_payload,
-                    "correlation_id": correlation_id,
-                }
-            )
-            return build_returns_series_response(
-                portfolio_returns=[
-                    ("2026-01-02", "0.0100"),
-                    ("2026-01-03", "-0.0050"),
-                    ("2026-01-04", "0.0040"),
-                ]
-            )
-
-    class _FakeLotusCoreClient:
-        def __init__(self) -> None:
-            self.position_calls: list[dict[str, object | None]] = []
-
-        async def get_position_analytics_timeseries(
-            self,
-            *,
-            portfolio_id: str,
-            request_payload: dict[str, object],
-            correlation_id: str | None,
-        ) -> dict[str, object]:
-            self.position_calls.append(
-                {
-                    "portfolio_id": portfolio_id,
-                    "request_payload": request_payload,
-                    "correlation_id": correlation_id,
-                }
-            )
-            return {
-                "rows": [
-                    {
-                        "security_id": "SEC_A",
-                        "valuation_date": "2026-01-02",
-                        "dimensions": {"sector": "TECH", "asset_class": "EQUITY"},
-                        "ending_market_value_portfolio_currency": "60",
-                    },
-                    {
-                        "security_id": "SEC_B",
-                        "valuation_date": "2026-01-02",
-                        "dimensions": {"sector": "HEALTH", "asset_class": "EQUITY"},
-                        "ending_market_value_portfolio_currency": "40",
-                    },
-                    {
-                        "security_id": "SEC_A",
-                        "valuation_date": "2026-01-03",
-                        "dimensions": {"sector": "TECH", "asset_class": "EQUITY"},
-                        "ending_market_value_portfolio_currency": "65",
-                    },
-                    {
-                        "security_id": "SEC_B",
-                        "valuation_date": "2026-01-03",
-                        "dimensions": {"sector": "HEALTH", "asset_class": "EQUITY"},
-                        "ending_market_value_portfolio_currency": "35",
-                    },
-                ],
-                "page": {"next_page_token": None},
-            }
-
-        async def get_instrument_enrichment(
-            self,
-            *,
-            security_ids: list[str],
-            correlation_id: str | None,
-        ) -> dict[str, object]:
-            return {"records": []}
-
-    performance_client = _FakeLotusPerformanceClient()
-    core_client = _FakeLotusCoreClient()
+    performance_client = build_stateful_attribution_returns_client()
+    core_client = RecordingHistoricalAttributionCoreClient()
     with override_app_runtime(
         lotus_performance_client=performance_client,
         lotus_core_client=core_client,
